+++ NEWS.md
@@ -29,5 +29,5 @@ timeout
 number of samples for deconvolution for quicker vignette/examples
 
 ## StatescopeR 0.99.32 (10-12-2025)
--   Adjusted BLADE code to pytorch (quicker version: https://github.com/tgac-vumc/Statescope/blob/master/src/BLADE_Deconvolution/BLADE.py)
+-   Adjusted to use BLADE code with pytorch (quicker version: https://github.com/tgac-vumc/Statescope/blob/master/src/BLADE_Deconvolution/BLADE.py)
 and made vignette less minimal